% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_Activity_API.R
\name{api_rate_limit}
\alias{api_rate_limit}
\title{Get rate limit of API}
\usage{
api_rate_limit(username = "", token = "")
}
\arguments{
\item{username}{(Optional) GitHub username for authentication and increasing rate limit.}

\item{token}{(Optional) GitHub access token for authentication and increasing rate limit.}
}
\value{
The string shows how much requests remaining based on limit.
}
\description{
This function gets rate limit information.
}
