% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_Activity_API.R
\name{format_response}
\alias{format_response}
\title{Format response}
\usage{
format_response(response, output_type)
}
\arguments{
\item{response}{Raw response gets from github_api().}

\item{output_type}{Indicate formatting type (Possible value: "list","dataframe","character","raw").}
}
\value{
Formatted response.
}
\description{
This function is a helper function for formatting response and return list/dataframe/character/raw response.
}
