% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_Activity_API.R
\name{github_api}
\alias{github_api}
\title{Send request to GitHub Activity API}
\usage{
github_api(path, per_page, page, username = "", token = "")
}
\arguments{
\item{path}{API path}

\item{per_page}{Results per page (max 100).}

\item{page}{Page number of the results to fetch.}

\item{username}{(Optional) GitHub username for authentication and increasing rate limit.}

\item{token}{(Optional) GitHub access token for authentication and increasing rate limit.}
}
\value{
Raw response
}
\description{
This function is a helper function for sending requests and return response.
}
