% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_Activity_API.R
\name{github_count_events_bydate}
\alias{github_count_events_bydate}
\title{Count events by date}
\usage{
github_count_events_bydate(
  range,
  max_event_num,
  owner = "",
  repo = "",
  organization = "",
  event_type = "all",
  return_type = "dataframe",
  username = "",
  token = ""
)
}
\arguments{
\item{range}{The range for retrieving events data (Possible value: "all public"/"network"/"organization"/"repository").}

\item{max_event_num}{The number of events data that user want.}

\item{owner}{(Optional) GitHub repository owner username.}

\item{repo}{(Optional) GitHub repository name.}

\item{organization}{(Optional) GitHub organization name.}

\item{event_type}{(Optional) The event type to count (default will count all types events).}

\item{return_type}{(Optional) Indicate return dataframe or plot (default: dataframe).}

\item{username}{(Optional) GitHub username for authentication and increasing rate limit.}

\item{token}{(Optional) GitHub access token for authentication and increasing rate limit.}
}
\value{
The dataframe/plot that contains the count information.
}
\description{
This function counts the number of events by date for a certain range and type.
}
