% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_Activity_API.R
\name{github_count_events_bytype}
\alias{github_count_events_bytype}
\title{Count events by type}
\usage{
github_count_events_bytype(
  range,
  max_event_num,
  owner = "",
  repo = "",
  organization = "",
  username = "",
  token = ""
)
}
\arguments{
\item{range}{The range for retrieving events data (Possible value: "all public"/"network"/"organization"/"repository").}

\item{max_event_num}{The number of events data that user want.}

\item{owner}{(Optional) GitHub repository owner username.}

\item{repo}{(Optional) GitHub repository name.}

\item{organization}{(Optional) GitHub organization name.}

\item{username}{(Optional) GitHub username for authentication and increasing rate limit.}

\item{token}{(Optional) GitHub access token for authentication and increasing rate limit.}
}
\value{
The dataframe that contains the count result.
}
\description{
This function counts the number of events by type for a certain range.
}
