% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_Activity_API.R
\name{github_get_public_events}
\alias{github_get_public_events}
\title{Get public events}
\usage{
github_get_public_events(
  per_page = 30,
  page = 1,
  output_type = "list",
  username = "",
  token = ""
)
}
\arguments{
\item{per_page}{(Optional) Results per page (max 100) (default 30).}

\item{page}{(Optional) Page number of the results to fetch (default 1).}

\item{output_type}{(Optional) Indicate formatting type (Possible value: "list","dataframe","character","raw") (default: "list").}

\item{username}{(Optional) GitHub username for authentication and increasing rate limit.}

\item{token}{(Optional) GitHub access token for authentication and increasing rate limit.}
}
\value{
Formatted response.
}
\description{
This function get public events data and return the data in user selected format.
}
