% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_Activity_API.R
\name{github_lastn_events_df}
\alias{github_lastn_events_df}
\title{Get last n events}
\usage{
github_lastn_events_df(
  range,
  max_event_num,
  owner = "",
  repo = "",
  organization = "",
  username = "",
  token = ""
)
}
\arguments{
\item{range}{The range for retrieving events data (Possible value: "all public"/"network"/"organization"/"repository").}

\item{max_event_num}{The number of events data that user want.}

\item{owner}{(Optional) GitHub repository owner username.}

\item{repo}{(Optional) GitHub repository name.}

\item{organization}{(Optional) GitHub organization name.}

\item{username}{(Optional) GitHub username for authentication and increasing rate limit.}

\item{token}{(Optional) GitHub access token for authentication and increasing rate limit.}
}
\value{
The dataframe that contains the retrieved events data.
}
\description{
This function get last n events data of a certain range and return the dataframe that contains the retrieved events data.
}
